Quicksort now uses unsigned ints to force empty items to appear higher. By promoting GetItemID to unsigned ints, we ensure that our 'no such item' value is actually higher than any valid item, specifically for the purposes of quicksort. It's a bit of a gross hack.

+++ sp/src/game/server/comrade_erika/baseinventory.cpp
@@ -499,14 +499,14 @@ void CBaseInventory::SwapItems(int itemindex1, int itemindex2)
 
 int CBaseInventory::QSPartition(int p, int r)
 {
-	int x = GetItemID(p);
+	unsigned int x = GetItemID(p);
 	int i = p - 1;
 	int j = r;
 
 	while (1)
 	{
-		do --j; while (GetItemID(j) > x);
-		do ++i; while (GetItemID(i) < x);
+		do --j; while ((unsigned int) GetItemID(j) > x);
+		do ++i; while ((unsigned int) GetItemID(i) < x);
 
 		if (i < j)
 			SwapItems(i, j);
